Check mistakes
