+++ project/src/components/map/map.tsx
@@ -1,0 +1,11 @@
+type MapProps = {
+  className: string;
+}
+
+function Map({className}: MapProps): JSX.Element {
+  return (
+    <section className={`${className} map`} />
+  );
+}
+
+export default Map;
