+++ project/src/components/place-card-mark/place-card-mark.tsx
@@ -1,0 +1,13 @@
+type PlaceCardMarkProps = {
+  className: string;
+}
+
+function PlaceCardMark({className}: PlaceCardMarkProps): JSX.Element {
+  return (
+    <div className={className}>
+      <span>Premium</span>
+    </div>
+  );
+}
+
+export default PlaceCardMark;
