+++ project/src/components/place-card/place-card.tsx
@@ -1,9 +1,14 @@
-function PlaceCard(): JSX.Element {
+import PlaceCardMark from '../place-card-mark/place-card-mark';
+import PlaceCardButton from '../place-card-button/place-card-button';
+
+type PlaceCardProps = {
+  className: string;
+}
+
+function PlaceCard({className}: PlaceCardProps): JSX.Element {
   return (
-    <article className="cities__place-card place-card">
-      <div className="place-card__mark">
-        <span>Premium</span>
-      </div>
+    <article className={`${className} place-card`}>
+      <PlaceCardMark className="place-card__mark" />
       <div className="cities__image-wrapper place-card__image-wrapper">
         <a href="#">
           <img className="place-card__image" src="img/apartment-01.jpg" width="260" height="200" alt="Place image" />
@@ -15,12 +20,7 @@ function PlaceCard(): JSX.Element {
             <b className="place-card__price-value">&euro;120</b>
             <span className="place-card__price-text">&#47;&nbsp;night</span>
           </div>
-          <button className="place-card__bookmark-button button" type="button">
-            <svg className="place-card__bookmark-icon" width="18" height="19">
-              <use xlinkHref="#icon-bookmark"></use>
-            </svg>
-            <span className="visually-hidden">To bookmarks</span>
-          </button>
+          <PlaceCardButton className="place-card__bookmark-button" />
         </div>
         <div className="place-card__rating rating">
           <div className="place-card__stars rating__stars">
