+++ project/src/const/const.tsx
@@ -1,0 +1,1 @@
+export const CITIES = ['Paris', 'Cologne', 'Brussels', 'Amsterdam', 'Hamburg', 'Dusseldorf'];
