+++ project/src/index.tsx
@@ -3,13 +3,11 @@ import ReactDOM from 'react-dom';
 import App from './components/app/app';
 
 const Setting = {
-  PLACES_COUNT: 327,
+  PLACES_COUNT: 312,
 };
 
 ReactDOM.render(
   <React.StrictMode>
-    <App
-      placesCount = {Setting.PLACES_COUNT}
-    />
+    <App placesCount={Setting.PLACES_COUNT} />
   </React.StrictMode>,
   document.getElementById('root'));
